+++ spec/unit/dream/scenarios/saving-types-that-dont-match-column-type.spec.ts
@@ -1,4 +1,6 @@
+import CheckConstraintViolation from '../../../../src/errors/db/CheckConstraintViolation.js'
 import DataTypeColumnTypeMismatch from '../../../../src/errors/db/DataTypeColumnTypeMismatch.js'
+import NotNullViolation from '../../../../src/errors/db/NotNullViolation.js'
 import { CalendarDate, DateTime } from '../../../../src/index.js'
 import { SpeciesTypesEnum } from '../../../../test-app/types/db.js'
 import { unpersistedFleshedOutModelForOpenapiTypeSpecs } from '../../../scaffold/fleshedOutModelForOpenapiTypeSpecs.js'
@@ -46,6 +48,29 @@ describe('saving types that don’t match the database column', () => {
       model.species = 'hello' as any as SpeciesTypesEnum
 
       await expect(model.save()).rejects.toThrow(DataTypeColumnTypeMismatch)
+    })
+  })
+
+  context('saving null to a not-null column', () => {
+    it('throws NotNullViolation', async () => {
+      const model = unpersistedFleshedOutModelForOpenapiTypeSpecs()
+      model.requiredCollarCountInt = null as unknown as number
+
+      await expect(model.save()).rejects.toThrow(NotNullViolation)
+    })
+  })
+
+  context('saving check constraint violation', () => {
+    it('throws CheckConstraintViolation', async () => {
+      const model = unpersistedFleshedOutModelForOpenapiTypeSpecs()
+
+      model.activateConstraint = true
+      model.myConstrainedField = null
+
+      await expect(model.save()).rejects.toThrow(CheckConstraintViolation)
+
+      model.myConstrainedField = 'hello'
+      await expect(model.save()).resolves.not.toThrow()
     })
   })
 })